+++ src/nexgen/beamlines/I19_2_gda_nxs.py
@@ -127,10 +127,6 @@ def tristan_writer(
         # Get correct start positions
         idx = [n for n, ax in enumerate(gonio_axes) if ax.name == k][0]
         gonio_axes[idx].start_pos = v[0]
-    # Change rotation axis direction for phi and omega because gda roation opposite to  epics
-    for ax in gonio_axes:
-        if ax.name in ["phi", "omega"]:
-            ax.vector = (-1, 0, 0)
     goniometer = Goniometer(gonio_axes, OSC)
 
     collection_summary_log(
@@ -217,10 +213,6 @@ def eiger_writer(
         # Get correct start positions
         idx = [n for n, ax in enumerate(gonio_axes) if ax.name == k][0]
         gonio_axes[idx].start_pos = v[0]
-    # Change rotation axis direction for phi and omega because gda roation opposite to  epics
-    for ax in gonio_axes:
-        if ax.name in ["phi", "omega"]:
-            ax.vector = (-1, 0, 0)
 
     # Get scan range array
     logger.info("Calculating scan range...")