+++ src/nexgen/tools/ED_tools.py
@@ -75,30 +75,32 @@ class SinglaMaster:
         if len(M) > 0:
             mask_path = [_loc for _loc in M if _loc.split("/")[-1] == "pixel_mask"]
             mask_applied_path = [_loc for _loc in M if "applied" in _loc]
-            if len(mask_applied_path) == 0:
-                return (None, self.__getitem__(mask_path[0])[()])
-            if len(mask_path) == 0:
-                return (self.__getitem__(mask_applied_path[0])[()], None)
-            return (
-                self.__getitem__(mask_applied_path[0])[()],
-                self.__getitem__(mask_path[0])[()],
+            mask_applied = (
+                False
+                if len(mask_applied_path) == 0
+                else self.__getitem__(mask_applied_path[0])[()]
             )
-        return (None, None)
+            mask = None if len(mask_path) == 0 else self.__getitem__(mask_path[0])[()]
+            return (mask_applied, mask)
+        return (False, None)
 
-    def get_flafield(self) -> Tuple[bool, ArrayLike]:
+    def get_flatfield(self) -> Tuple[bool, ArrayLike]:
         F = [obj for obj in self.walk if "flatfield" in obj]
         if len(F) > 0:
             flatfield_path = [_loc for _loc in F if _loc.split("/")[-1] == "flatfield"]
             flatfield_applied_path = [_loc for _loc in F if "applied" in _loc]
-            if len(flatfield_applied_path) == 0:
-                return (None, self.__getitem__(flatfield_path[0])[()])
-            if len(flatfield_path) == 0:
-                return (self.__getitem__(flatfield_applied_path[0])[()], None)
-            return (
-                self.__getitem__(flatfield_applied_path[0])[()],
-                self.__getitem__(flatfield_path[0])[()],
+            flatfield_applied = (
+                False
+                if len(flatfield_applied_path) == 0
+                else self.__getitem__(flatfield_applied_path[0])[()]
             )
-        return (None, None)
+            flatfield = (
+                None
+                if len(flatfield_path) == 0
+                else self.__getitem__(flatfield_path[0])[()]
+            )
+            return (flatfield_applied, flatfield)
+        return (False, None)
 
     def get_bit_bepth_readout(self) -> int:
         _loc = [obj for obj in self.walk if "bit_depth_readout" in obj]
@@ -228,7 +230,7 @@ def extract_detector_info_from_master(master: Path | str) -> Dict[str, Any]:
         MASK = singla.get_mask()
         if MASK[1] is not None:
             logger.info("Pixel_mask has been found in master file.")
-        FF = singla.get_flafield()
+        FF = singla.get_flatfield()
         if FF[1] is not None:
             logger.info("Flatfield has been found in master file.")
         D["pixel_mask"] = MASK[1]