+++ tests/nxs_write/conftest.py
@@ -17,6 +17,17 @@ from nexgen.nxs_utils import (
 )
 from nexgen.nxs_write.NXmxWriter import NXmxFileWriter
 
+dummy_config = """{
+    "nimages": 90,
+    "ntrigger": 1,
+    "omega_increment": 1.0,
+    "omega_start": 0.0,
+    "phi_increment": 0.0,
+    "phi_start": 0.0,
+    "two_theta_start": 10.,
+    "two_theta_increment": 0.0,
+    }"""
+
 
 @pytest.fixture
 def mock_goniometer() -> Goniometer:
@@ -72,6 +83,25 @@ def mock_eiger() -> Detector:
         0.01,
         [(1, 0, 0), (0, 1, 0)],
     )
+
+
+@pytest.fixture
+def dummy_eiger_meta_file():
+    test_hdf_file = tempfile.TemporaryFile()
+    test_meta_file = h5py.File(test_hdf_file, "w")
+    test_meta_file["config"] = dummy_config
+    test_meta_file["_dectris/nimages"] = np.array([90])
+    test_meta_file["_dectris/ntrigger"] = np.array([1])
+    test_meta_file["_dectris/wavelength"] = np.array([0.6])
+    test_meta_file["_dectris/x_pixels_in_detector"] = np.array([3180])
+    test_meta_file["_dectris/y_pixels_in_detector"] = np.array([3262])
+    test_meta_file["_dectris/detector_distance"] = np.array([0.19])
+    test_meta_file["_dectris/bit_depth_readout"] = np.array([32])
+    test_meta_file["flatfield"] = np.array([[0, 0, 0]])
+    test_meta_file["_dectris/software_version"] = np.string_("0.0.0")
+    test_meta_file["mask"] = np.array([[0, 1, 1], [1, 0, 0]])
+    test_meta_file["_dectris/pixel_mask_applied"] = False
+    yield test_meta_file
 
 
 @pytest.fixture